Tweak build and rhosp scripts to support vol backed images

+++ setup-kvm-and-openstack/openstack-scripts/openstack-env-config.sh
@@ -145,8 +145,9 @@ chown ${USERNAME}:${USERNAME} /home/${USERNAME}/keystonerc_${USERNAME}
 }
 
 create-images() {
-  # Logging everything except the image creation
+  # image creation
   cmd echo "INFO: Starting function 'create-images'"
+  source /root/keystonerc_${USERNAME}
 
   if ! glance image-list | grep image-base-src
   then
@@ -159,6 +160,28 @@ create-images() {
       --file ${OPENSHIFT_IMAGE_PATH} \
       image-base-src
   fi
+	# wait for image to become active
+	echo -en "\nWaiting for image to create"
+  counter=0
+	while :
+	do
+    counter=$(( $counter + 1 ))
+		echo -n "."
+		sleep 1
+	  if openstack image show image-base-src -f value -c status | grep -q active
+    then
+      break
+    fi
+    if [ $counter -gt $TIMEOUT ]
+    then
+      echo ERROR: something went wrong - check console
+      exit 1
+    elif [ $counter -eq $TIMEOUT_WARN ]
+    then
+      echo -n "WARN: this is taking longer than expected"
+    fi
+	done
+	echo ""
   if ! glance image-list | grep ${OPENSHIFT_VM_NAME}
   then
     cmd openstack volume create \
@@ -173,8 +196,39 @@ create-images() {
       --property hw_disk_bus=scsi \
       --min-disk 10 \
       ${OPENSHIFT_VM_NAME}
+    echo -en "\nWaiting for volume to create"
+    counter=0
+    while :
+    do
+      counter=$(( $counter + 1 ))
+      echo -n "."
+      sleep 1
+      if openstack volume show ${OPENSHIFT_VM_NAME} -f value -c status | grep -q available
+      then
+        break
+      fi
+      if [ $counter -gt $TIMEOUT ]
+      then
+        echo ERROR: something went wrong - check console
+        exit 1
+      elif [ $counter -eq $TIMEOUT_WARN ]
+      then
+        echo -n "WARN: this is taking longer than expected"
+      fi
+    done
+    echo ""
     VOLUME_ID=$(openstack volume show ${OPENSHIFT_VM_NAME} -f value -c id)
     IMAGE_ID=$(openstack image show ${OPENSHIFT_VM_NAME} -f value -c id)
+    if openstack image list | grep -qi error
+    then
+      echo "ERROR: Image creation failed"
+      exit 1
+    fi
+    if openstack volume list | grep -qi error
+    then
+      echo "ERROR: Volume creation failed"
+      exit 1
+    fi
     cmd glance location-add ${IMAGE_ID} --url cinder://${VOLUME_ID}
     cmd virt-sparsify /dev/cinder-volumes/volume-${VOLUME_ID} --in-place
     cmd openstack image delete image-base-src
@@ -256,9 +310,9 @@ build-instances() {
     then
       echo ERROR: something went wrong - check console
       exit 1
-    elif [ $counter -gt $TIMEOUT_WARN ]
+    elif [ $counter -eq $TIMEOUT_WARN ]
     then
-      echo WARN: this is taking longer than expected
+      echo -n "WARN: this is taking longer than expected"
     fi
 	done
 	echo ""
@@ -289,11 +343,11 @@ verify-networking() {
     fi
     if [ $counter -gt $TIMEOUT ]
     then
-      echo ERROR: something went wrong - check console
+      echo "ERROR: something went wrong - check console"
       exit 1
-    elif [ $counter -gt $TIMEOUT_WARN ]
+    elif [ $counter -eq $TIMEOUT_WARN ]
     then
-      echo WARN: this is taking longer than expected
+      echo -n "WARN: this is taking longer than expected"
     fi
 	done
 	echo ""
@@ -319,14 +373,14 @@ cleanup() {
     fi
     if [ $counter -gt $TIMEOUT ]
     then
-      echo ERROR: something went wrong - check console
+      echo "ERROR: something went wrong - check console"
       exit 1
-    elif [ $counter -gt $TIMEOUT_WARN ]
+    elif [ $counter -eq $TIMEOUT_WARN ]
     then
-      echo WARN: this is taking longer than expected
+      echo -n "WARN: this is taking longer than expected"
     fi
   done
-   echo ""
+  echo ""
   echo -n "Waiting for openshift-base-volume to be deleted"
   counter=0
   while :
@@ -340,11 +394,11 @@ cleanup() {
     fi
     if [ $counter -gt $TIMEOUT ]
     then
-      echo ERROR: something went wrong - check console
+      echo "ERROR: something went wrong - check console"
       exit 1
-    elif [ $counter -gt $TIMEOUT_WARN ]
+    elif [ $counter -eq $TIMEOUT_WARN ]
     then
-      echo WARN: this is taking longer than expected
+      echo -n "WARN: this is taking longer than expected"
     fi
   done
   echo ""
@@ -359,6 +413,7 @@ post-install-config 2>&1 | tee -a ${LOGFILE}
 post-install-admin-tasks 2>&1 | tee -a ${LOGFILE}
 create-images 2>&1 | tee -a ${LOGFILE}
 post-install-user-tasks 2>&1 | tee -a ${LOGFILE}
+# Commenting out the following functions intentionally to avoid local image cache creation in /var/lib/nova/instances
 #build-instances 2>&1 | tee -a ${LOGFILE}
 #source /root/keystonerc_${USERNAME}
 #if nova list | grep ERROR
